add GetSignatureInfo and deploy contracts on new addresses
