+++ on-chain/scripts/mumbai-dev/set-dao.js
@@ -4,20 +4,19 @@ const hre = require("hardhat");
 async function main() {
 
  // todo: change address
-  const oracleDAOContractAddress = "0xe3262c0848b0cc5cd43df7139103f1fbf26558cc";
+  const oracleDAOContractAddress = "0x9208C2B417Ec2699454843A06A5E49fA6dd88422";
 
   // todo: update addresses
   const addressList = [
-    "0x800210CfB747992790245eA878D32F188d01a03A",
-    "0x05856015d07F3E24936B7D20cB3CcfCa3D34B41d",
-    "0x6f2B76024196e82D81c8bC5eDe7cff0B0276c9C1"
+    "0x6d2e5279b106843f6E924194401B50e6e27FE12a",
+    "0xbE14Eb1ffcA54861D3081560110a45F4A1A9e9c5",
+    "0xeA2bf08288bbfB0d3DBf534f35af32bF2c6E5e45"
   ];
 
   const [signer] = await ethers.getSigners();
 
   const contract = await hre.ethers.getContractFactory("FilswanOracle");
   const daoOracleInstance = await contract.attach(oracleDAOContractAddress);
-
 
   const tx = await daoOracleInstance.connect(signer).setDAOUsers(addressList);
   await tx.wait();
